+++ packages/engine-server/src/topics/site.ts
@@ -181,7 +181,7 @@ export class SiteUtils {
     // async pass to process all notes
     const domainsAndhiearchiesToPublish = await Promise.all(
       siteHierarchies.map(async (domain, idx) => {
-        const out = await SiteUtils.filterByHiearchy({
+        const out = await SiteUtils.filterByHierarchy({
           domain,
           config,
           engine,
@@ -203,7 +203,7 @@ export class SiteUtils {
       domains.push(domain);
       hiearchiesToPublish.push(notes);
     });
-    // if single hiearchy, domain includes all immediate children
+    // if single hierarchy, domain includes all immediate children
     if (
       !opts.noExpandSingleDomain &&
       siteHierarchies.length === 1 &&
@@ -233,9 +233,9 @@ export class SiteUtils {
   }
 
   /**
-   * Filter notes to be published using hiearchy
+   * Filter notes to be published using hierarchy
    */
-  static async filterByHiearchy(opts: {
+  static async filterByHierarchy(opts: {
     domain: string;
     config: IntermediateDendronConfig;
     engine: DEngineClient;
@@ -243,12 +243,12 @@ export class SiteUtils {
   }): Promise<{ notes: NotePropsDict; domain: NoteProps } | undefined> {
     const { domain, engine, navOrder, config } = opts;
     const logger = createLogger(LOGGER_NAME);
-    logger.info({ ctx: "filterByHiearchy:enter", domain, config });
+    logger.info({ ctx: "filterByHierarchy:enter", domain, config });
     const hConfig = this.getConfigForHierarchy({
       config,
       noteOrName: domain,
     });
-    const notesForHiearchy = _.clone(engine.notes);
+    const notesForHierarchy = _.clone(engine.notes);
 
     // get the domain note
     const notes = NoteUtils.getNotesByFnameFromEngine({
@@ -256,7 +256,7 @@ export class SiteUtils {
       engine,
     });
     logger.info({
-      ctx: "filterByHiearchy:candidates",
+      ctx: "filterByHierarchy:candidates",
       domain,
       hConfig,
       notes: notes.map((ent) => ent.id),
@@ -274,12 +274,12 @@ export class SiteUtils {
         config,
         fname: domain,
         noteCandidates: notes,
-        noteDict: notesForHiearchy,
+        noteDict: notesForHierarchy,
       });
       // no note found
     } else if (notes.length < 1) {
       logger.error({
-        ctx: "filterByHiearchy",
+        ctx: "filterByHierarchy",
         msg: "note not found",
         domain,
       });
@@ -313,7 +313,7 @@ export class SiteUtils {
     }
 
     logger.info({
-      ctx: "filterByHiearchy",
+      ctx: "filterByHierarchy",
       domainNote: NoteUtils.toNoteLoc(domainNote),
     });
 
@@ -326,7 +326,7 @@ export class SiteUtils {
       let note = processQ.pop() as NoteProps;
 
       logger.debug({
-        ctx: "filterByHiearchy",
+        ctx: "filterByHierarchy",
         maybeNote: NoteUtils.toNoteLoc(note),
       });
 
@@ -346,7 +346,7 @@ export class SiteUtils {
       let children = HierarchyUtils.getChildren({
         skipLevels: siteFM.skipLevels || 0,
         note,
-        notes: notesForHiearchy,
+        notes: notesForHierarchy,
       });
       if (siteFM.skipLevels && siteFM.skipLevels > 0) {
         note.children = children.map((ent) => ent.id);
@@ -364,7 +364,7 @@ export class SiteUtils {
         })
       );
       logger.debug({
-        ctx: "filterByHiearchy:post-filter-children",
+        ctx: "filterByHierarchy:post-filter-children",
         note: NoteUtils.toNoteLoc(note),
         children: children.map((ent) => ent.id),
       });
@@ -577,7 +577,7 @@ export class SiteUtils {
       const logger = createLogger(LOGGER_NAME);
       if (maybeDomainNotes.length < 1) {
         logger.error({
-          ctx: "filterByHiearchy",
+          ctx: "filterByHierarchy",
           msg: "dup-resolution: no note found",
           vault,
         });
@@ -606,7 +606,7 @@ export class SiteUtils {
     });
     const logger = createLogger(LOGGER_NAME);
     logger.info({
-      ctx: "filterByHiearchy",
+      ctx: "filterByHierarchy",
       msg: "dup-resolution: resolving dup",
       parent: domainNote.id,
       children: domainNote.children,